requirements.txt
functorch
matplotlib
torch-summary
tqdm
pyyaml
easydict
